requirements.txt
streamlit
groq
faiss-cpu
sentence-transformers
langchain-text-splitters
python-dotenv
numpy